perf: initialize ajv validator only once
This is a small performance improvement in which an ajv validator for the
ng-package JSON schema is only compiled once. Additionally, the schema
validation logic is refactored into its own file.
